Refactor pricing section into a separate component
✅✅ Extract pricing section from main page into a dedicated PricingSection component
✅ Remove inline pricing plans from page.tsx
- Maintain existing styling and animation effects
- Improve code organization and modularity
