+++ src/components/footer.tsx
@@ -1,83 +1,69 @@
 import Link from "next/link";
 
+const footerLinks = [
+  {
+    title: "Quick Links",
+    links: [
+      { name: "Features", href: "#features" },
+      { name: "Pricing", href: "#pricing" },
+    ],
+  },
+  {
+    title: "Legal",
+    links: [
+      { name: "Terms of Service", href: "/terms" },
+      { name: "Privacy Policy", href: "/privacy" },
+    ],
+  },
+];
+
 export default function Footer() {
   return (
-    <footer className="flex flex-col items-center justify-center gap-6 text-center">
-      <div className="container mx-auto px-4">
-        <div className="grid grid-cols-1 md:grid-cols-4 gap-8">
+    <footer className="bgParticles w-full text-white py-10">
+      <div className="max-w-5xl mx-auto px-6">
+        <div className="flex flex-col lg:flex-row justify-between gap-6">
           <div>
-            <h3 className="text-lg font-semibold mb-4">Quinx</h3>
-            <p className="text-gray-400">
-              Building the future of Discord bots, <br />
-              one click at a time.
+            <h3 className="text-lg font-semibold">Quinx</h3>
+            <p className="text-gray-400 text-sm mt-2">
+              Building the future of Discord bots, one click <br /> at a time.
             </p>
           </div>
-          <div>
-            <h4 className="text-lg font-semibold mb-4">Quick Links</h4>
-            <ul className="space-y-2">
-              <li>
-                <Link
-                  href="#features"
-                  className="text-gray-400 hover:text-white transition-colors"
-                >
-                  Features
-                </Link>
-              </li>
-              <li>
-                <Link
-                  href="#pricing"
-                  className="text-gray-400 hover:text-white transition-colors"
-                >
-                  Pricing
-                </Link>
-              </li>
-            </ul>
+          <div className="grid grid-cols-2 gap-6">
+            {footerLinks.map((section) => (
+              <div key={section.title}>
+                <h4 className="text-lg font-semibold">{section.title}</h4>
+                <ul className="mt-2 space-y-1 text-sm">
+                  {section.links.map((link) => (
+                    <li key={link.name}>
+                      <Link
+                        href={link.href}
+                        className="text-gray-400 hover:text-white"
+                      >
+                        {link.name}
+                      </Link>
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            ))}
           </div>
-          <div>
-            <h4 className="text-lg font-semibold mb-4">Legal</h4>
-            <ul className="space-y-2">
+          {/*  <div>
+            <h4 className="text-lg font-semibold">Connect</h4>
+            <ul className="mt-2 space-y-1 text-sm">
               <li>
-                <Link
-                  href="/terms"
-                  className="text-gray-400 hover:text-white transition-colors"
-                >
-                  Terms of Service
-                </Link>
-              </li>
-              <li>
-                <Link
-                  href="/privacy"
-                  className="text-gray-400 hover:text-white transition-colors"
-                >
-                  Privacy Policy
-                </Link>
-              </li>
-            </ul>
-          </div>
-          <div>
-            <h4 className="text-lg font-semibold mb-4">Connect</h4>
-            <ul className="space-y-2">
-              <li>
-                <a
-                  href="#"
-                  className="text-gray-400 hover:text-white transition-colors"
-                >
+                <a href="#" className="text-gray-400 hover:text-white">
                   Twitter
                 </a>
               </li>
-
               <li>
-                <a
-                  href="#"
-                  className="text-gray-400 hover:text-white transition-colors"
-                >
+                <a href="#" className="text-gray-400 hover:text-white">
                   Discord
                 </a>
               </li>
             </ul>
-          </div>
+          </div> */}
         </div>
-        <div className="mt-8 pt-8 border-t border-gray-800 text-center text-gray-400">
+        <div className="mt-8 border-t border-gray-800 pt-4 text-gray-400 text-sm text-center">
           <p>&copy; {new Date().getFullYear()} Quinx. All rights reserved.</p>
         </div>
       </div>
